Move harmony patch files to new dir.
