+++ Source/ResearchTree/Patches/MainTabWindow_Research_DoBeginResearch.cs
@@ -7,7 +7,7 @@ using Verse;
 
 namespace FluffyResearchTree;
 
-[HarmonyPatch(typeof(MainTabWindow_Research), "DoBeginResearch")]
+[HarmonyPatch(typeof(MainTabWindow_Research), nameof(MainTabWindow_Research.DoBeginResearch))]
 public class MainTabWindow_Research_DoBeginResearch
 {
     private static void Prefix(ResearchProjectDef projectToStart)
